Video -> canvas streaming util separated out.
